chrome/test/data/xr/e2e_test_files/html/test_webxr_gamepad_support.html

<!doctype html>
<!--
Tests that a Gamepad API gamepad is or is not returned under different
circumstances.
-->
<html>
  <head>
    <link rel="stylesheet" type="text/css" href="../resources/webxr_e2e.css">
  </head>
  <body>
    <canvas id="webgl-canvas"></canvas>
    <script src="../../../../../../third_party/blink/web_tests/resources/testharness.js"></script>
    <script src="../resources/webxr_e2e.js"></script>
    <script src="../resources/webxr_boilerplate.js"></script>
    <script>
      // We apparently need to register a listener, otherwise all gamepads are
      // always null.
      window.addEventListener("gamepadconnected", function(e) {});

      // TODO(https://crbug.com/961774): On some Windows machines, a phantom
      // gamepad shows up in the navigator.getGamepads() array once a WebXR
      // gamepad sends any input. Until that's fixed, filter out such devices
      // when checking if the number of gamepads on the navigator array matches
      // expectations.
      function isPhantomGamepad(gamepad) {
        return gamepad.id == "Unknown Gamepad (Vendor: 0000 Product: 0000)";
      }

      function assertNumNavigatorGamepadsMatchesExpectation(numExpectedGamepads) {
        let numGamepads = 0;
        for (gamepad of navigator.getGamepads()) {
          if ((gamepad !== null) && !isPhantomGamepad(gamepad)) {
            numGamepads++;
          }
        }
        assert_equals(numGamepads, numExpectedGamepads,
            "Number of returned gamepads matches expectation");
      }

      function assertNumInputSourceGamepadsMatchesExpectation(numExpectedGamepads) {
        let currentSession = sessionInfos[sessionTypes.IMMERSIVE].currentSession;
        let numGamepads = 0;
        for (source of currentSession.getInputSources()) {
          if (source.gamepad !== null) {
            numGamepads++;
          }
        }
        assert_equals(numGamepads, numExpectedGamepads,
            "Number of XRInputSources with Gamepads on XRSession matches expectation");
      }

      function assertNumInputSourcesMatchesExpectation(numExpectedSources) {
        let currentSession = sessionInfos[sessionTypes.IMMERSIVE].currentSession;
        assert_equals(currentSession.getInputSources().length, numExpectedSources,
            "Number of XRInputSources on XRSession matches expectation");
      }

      function validateInputSourceHasNoGamepad() {
        assert_equals(onImmersiveXRFrameCallback, null);
        onImmersiveXRFrameCallback = function(session) {
          // Clear the callback first in case any of the following asserts fail.
          // If any asserts trigger, we don't run any code that follows, so if
          // we did this as the last step of the callback and any of the asserts
          // failed, we would be continually called, which just makes it harder
          // to debug.
          onImmersiveXRFrameCallback = null;

          // We don't expect to have attached any non-vr gamepads, and vr
          // gamepads shouldn't show up in navigator.getGamepads()
          assertNumNavigatorGamepadsMatchesExpectation(0);

          // There should only be one input source, but there should not be
          // enough data to make a gamepad.
          assertNumInputSourcesMatchesExpectation(1);
          assertNumInputSourceGamepadsMatchesExpectation(0);
          finishJavaScriptStep();
        }
      }

      function validateFirstGamepad(validationFunction) {
        assert_equals(onImmersiveXRFrameCallback, null);
        onImmersiveXRFrameCallback = function(session) {
          // Clear the callback first in case any of the following asserts fail.
          // If any asserts trigger, we don't run any code that follows, so if
          // we did this as the last step of the callback and any of the asserts
          // failed, we would be continually called, which just makes it harder
          // to debug.
          onImmersiveXRFrameCallback = null;

          // We don't expect to have attached any non-vr gamepads, and vr
          // gamepads shouldn't show up in navigator.getGamepads()
          assertNumNavigatorGamepadsMatchesExpectation(0);

          // There should only be one input source and it should have a gamepad.
          assertNumInputSourcesMatchesExpectation(1);
          assertNumInputSourceGamepadsMatchesExpectation(1);

          let gamepad = session.getInputSources()[0].gamepad;
          if (validationFunction) {
            validationFunction(gamepad);
          }

          finishJavaScriptStep();
        }
      }

      function validateMapping(expected_mapping) {
        let validationFunction = function(gamepad) {
          assert_equals(gamepad.mapping, expected_mapping);
        }

        validateFirstGamepad(validationFunction);
      }

      function validateButtonNotPressed(button_index) {
        let validationFunction = function(gamepad) {
          assert_less_than(button_index, gamepad.buttons.length,
            "Verify that we have at least as many buttons as requested");
          assert_false(gamepad.buttons[button_index].pressed);
        }

        validateFirstGamepad(validationFunction);
      }

      function validateButtonPressed(button_index) {
        let validationFunction = function(gamepad) {
          assert_less_than(button_index, gamepad.buttons.length,
            "Verify that we have at least as many buttons as requested");
          assert_true(gamepad.buttons[button_index].pressed);
        }

        validateFirstGamepad(validationFunction);
      }

      function validateButtonTouched(button_index) {
        let validationFunction = function(gamepad) {
          assert_less_than(button_index, gamepad.buttons.length,
            "Verify that we have at least as many buttons as requested");
          assert_true(gamepad.buttons[button_index].touched);
        }

        validateFirstGamepad(validationFunction);
      }

      function validateAxesValues(axes_pair_index, x_axis_value, y_axis_value) {
        let validationFunction = function(gamepad) {
          let epsilon = 0.001;
          let x_index = (2 * axes_pair_index);
          let y_index = (2 * axes_pair_index) + 1;

          assert_less_than(y_index, gamepad.axes.length,
            "Verify that we have at least as many axes as requested");
          assert_true(Math.abs(gamepad.axes[x_index] - x_axis_value) < epsilon,
            "X approximately equals expected");
          assert_true(Math.abs(gamepad.axes[y_index] - y_axis_value) < epsilon,
            "Y approximately equals expected");
        }

        validateFirstGamepad(validationFunction);
      }
    </script>
  </body>
</html>
